+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 36.3.18
+
+### Patch Changes
+
+- Updated dependencies
+
+## 36.3.17
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 36.3.16
 
 ### Patch Changes